+++ RELEASE-NOTES.md
@@ -52,6 +52,7 @@
 - The scheduler flag `-zk_use_curator` has been deprecated. If you have never set the flag and are
   upgrading you should take care as described in the [note](#zk_use_curator_upgrade) above.
 - The `key` argument of `getJobUpdateDetails` has been deprecated. Use the `query` argument instead.
+- The --release-threshold option on `aurora job restart` has been removed.
 
 0.15.0
 ======